requirements.txt
langchain
langchain-openai
langchain-community
streamlit
pandas
pdfplumber
tabula-py
# yaml